+++ common/src/main/java/dev/gimme/adventurezones/application/ChunkHandler.java
@@ -1,6 +1,8 @@
 package dev.gimme.adventurezones.application;
 
 import dev.gimme.adventurezones.domain.AdventureZones;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.chunk.LevelChunk;
 import org.jetbrains.annotations.NotNull;
 
@@ -20,5 +22,11 @@ public class ChunkHandler {
     public void onChunkUnload(@NotNull LevelChunk chunk) {
         if (chunk.getLevel().isClientSide()) return;
         adventureZones.unloadChunk(chunk.getPos());
+    }
+
+    public void onBlockPlace(@NotNull LevelChunk chunk, @NotNull BlockPos pos, @NotNull BlockState state) {
+        if (chunk.getLevel().isClientSide()) return;
+
+        adventureZones.onPlayerPlaceBlock(chunk, pos, state);
     }
 }