+++ neoforge/src/main/java/dev/gimme/adventurezones/infrastructure/ChunkListener.java
@@ -3,6 +3,7 @@ package dev.gimme.adventurezones.infrastructure;
 import dev.gimme.adventurezones.application.ChunkHandler;
 import net.minecraft.world.level.chunk.LevelChunk;
 import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.neoforge.event.level.BlockEvent;
 import net.neoforged.neoforge.event.level.ChunkEvent;
 import org.jetbrains.annotations.NotNull;
 
@@ -24,5 +25,19 @@ public class ChunkListener {
     public void onChunkUnload(ChunkEvent.Unload event) {
         if (!(event.getChunk() instanceof LevelChunk chunk)) return;
         chunkHandler.onChunkUnload(chunk);
+    }
+
+    @SubscribeEvent
+    public void onBlockPlace(BlockEvent.EntityPlaceEvent event) {
+        if (!(event.getLevel().getChunk(event.getPos()) instanceof LevelChunk chunk)) return;
+
+        chunkHandler.onBlockPlace(chunk, event.getPos(), event.getPlacedBlock());
+    }
+
+    @SubscribeEvent
+    public void onMultiBlockPlace(BlockEvent.EntityMultiPlaceEvent event) {
+        if (!(event.getLevel().getChunk(event.getPos()) instanceof LevelChunk chunk)) return;
+
+        chunkHandler.onBlockPlace(chunk, event.getPos(), event.getPlacedBlock());
     }
 }